convert numpy float32 into python float

+++ src/models/evaluate.py
@@ -3,8 +3,8 @@ from sklearn.neighbors import NearestNeighbors
 
 def evaluate_model(nn: NearestNeighbors, X: np.ndarray) -> tuple[float, float]:
     distances, _ = nn.kneighbors(X)
-    mean_dist = distances.mean()
-    median_dist = np.median(distances)
+    mean_dist = float(distances.mean())
+    median_dist = float(np.median(distances))
 
     return mean_dist, median_dist
